Chore(auth) : INTEG-247 refactoring code about slo
