+++ session/src/main/java/org/entcore/session/AuthManager.java
@@ -730,7 +730,7 @@ public class AuthManager extends BusModBase implements Handler<Message<JsonObjec
 					final String userId = ar.result().getString("userId");
 					eb.send("saml", new JsonObject().put("action", "soap-slo").put("sessionId", sessionId).put("userId", userId));
 				}
-				if (getOrElse(config.getBoolean("slo-oidc-backchannel-logout"), false)) {
+				if (getOrElse(config.getBoolean("slo-oidc-backchannel-logout"), true)) {
 					final String userId = ar.result().getString("userId");
 					eb.send("openid",
 							new JsonObject().put("action", "oidc-slo").put("userId", userId).put("sessionId",
